+++ src/main/clojure/com/github/clojure_repl/intellij/action/eval.clj
@@ -56,7 +56,7 @@
                root-zloc (z/of-string text)
                zloc (parser/find-form-at-pos root-zloc (inc row) col)
                code (z/string zloc)
-               ns (some-> (parser/find-namespace root-zloc) z/string)]
+               ns (some-> (parser/find-namespace root-zloc) z/string parser/remove-metadata)]
            (nrepl/eval {:project (.getProject editor) :code code :ns ns})))
        (fn [response]
          (string/join "\n" (:value response)))))))
@@ -72,7 +72,7 @@
            zloc (-> (parser/find-form-at-pos root-zloc (inc row) col)
                     parser/to-top)
            code (z/string zloc)
-           ns (some-> (parser/find-namespace root-zloc) z/string)]
+           ns (some-> (parser/find-namespace root-zloc) z/string parser/remove-metadata)]
        (nrepl/eval {:project (.getProject editor) :code code :ns ns})))
    (fn [response]
      (string/join "\n" (:value response)))))
@@ -90,7 +90,7 @@
      (let [text (.getText (.getDocument editor))
            root-zloc (z/of-string text)
            zloc (parser/find-namespace root-zloc)
-           namespace (z/string zloc)]
+           namespace (parser/remove-metadata (z/string zloc))]
        (nrepl/eval {:project (.getProject editor) :code (format "(in-ns '%s)" namespace) :ns namespace})))
    (fn [response]
      (string/join "\n" (:value response)))))