+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Fix ignore namespace metadata in REPL 
+
 ## 1.4.1
 
 - Fix REPL input parse of functions with `>` char. #79